Turning point!
This version has a few bugs. ExpandAreas get their children confused.
Because of these technical problems and because the current viewing of
elements in a set is bad, this version is being committed for a rollback
option, and a new version shall be created where the list of elements
within a set is in a new Activity